requirements.txt
pandas>=2.0.0
xlrd>=2.0.1
python-docx>=1.1.0
docxtpl>=0.16.7
weasyprint>=60.0
jinja2>=3.1.2
python-dateutil>=2.8.2
openpyxl>=3.1.2
unidecode>=1.3.6
pyexcel==0.7.0
pyexcel-xls==0.7.1
pyexcel-xlsx==0.6.1
pdfplumber>=0.5.28
PyMuPDF>=1.22.5
fastapi
uvicorn[standard]
app
python-multipart
